commit the Order-Details Page changes and some other changes

+++ Models/orders.cs
@@ -14,10 +14,10 @@ namespace Spendwise_WebApp.Models
         public required string CompanyName { get; set; }
         public int? CompanyId { get; set; }
         public string AdditionalPackageItemIds { get; set; } 
-        public required double NetAmount { get; set; }
-        public required double VatAmount { get; set; }
-        public required double TotalAmount { get; set; }
-        public required double AmountDue { get; set; }
+        public required decimal NetAmount { get; set; }
+        public required decimal VatAmount { get; set; }
+        public required decimal TotalAmount { get; set; }
+        public required decimal AmountDue { get; set; }
         public bool IsOrderComplete { get; set; }
         public DateTime? InvoicedDate { get; set; }
         public bool IsPaymentComplete { get; set; }
